#71: Create checkpointing module
